backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# ML dependencies
scikit-learn==1.4.0
joblib==1.3.2
pandas==2.2.0
numpy==1.26.3

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Data processing (for training scripts)
matplotlib==3.8.2
seaborn==0.13.1

# Security (optional but recommended)
cryptography==42.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
